requirements.txt
pyarrow
oracledb
python-dotenv
adbc-driver-postgresql
pandas
